refactor(namespace): split mount propagation module
Separate peer-group management, per-mount propagation state, recursive type
changes, topology event handling, and white-box tests into focused modules
behind the existing facade.

Encapsulate fallible propagation snapshots and peer registry transaction
updates so recursive changes preserve allocation failure atomicity, lifecycle
lock ordering, Weak mount ownership, and lock-external group release semantics.

Update the stale white-box propagation test to exercise the current prepare,
topology publication, and commit sequence without changing production behavior
or test expectations.
